Delete Account-function
Added a way for the user to delete their account in Profile.
It will prompt user to enter their password then send a confirmation-email, after clicking the confirmation-link the account will be terminated and a success-message will be displayed.
Subscribed users gets a warning displayed in the Account Deletion-form stating that the subscription will no longer be available after account-deletion.